backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langgraph
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
tavily-python
pypdf
